imgbase: Add experimental support for syslinux
Previously imgbased used BLS config files to install new boot entries,
the problem with this approach is that BLS does not allow to set a
default entry.
Now the syslinux configuration file format is used to specify the
bootloader configuration. This format allows to set the default boot
entry to be booted.

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>