imgbase: Add init-from
Allow creating a snapshotable layout from an existing logical volume.
This works by taking a snapshot of the original LV, this snapshot is then
following the imgbased naming guidelines (which are beeing rethought). This
allows imagbased to discover this stream of images and handle them correctly.
To make this work, imgbase is now using tags to discover a VG which contains
a stream of images.

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>