test: Refactor testing to use FakeLVM
Previously the testing required to be run in a VM with
a fully functional LVM available.

TO ease testing now a fake LVM (FakeLVM) class is used to
patch out the imgbased.lvm.LVM class.
It mimics a bug free LVM behavior and can be used to validate
the high-level logic of imgbased.

Benefits are:
- Does not require root privileges
- Independent of LVM presence and layout
- Pure pythonic
- Fast

Change-Id: I40dba8ad07203bdc5618a63ea81d259ba05003a1
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>