osupdater: Migrate config from previous layer, not from previous LV
Previously the osupdater tried to migrate the configs from a previous LV to
the new layer. And a previous LV could be a previous base, and this is the wrong
logic.

Our assumption is that every previous layer (not LV) has the correct /etc
contents, thus we need to migrate the configs from there.

Change-Id: Ie21d43831afdcdd708190ad99e3e5339673e0725
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>