naming: Rework to have an rpm compatible scheme
Previously the bases and layers were organized my
modifying the version of an image, i.e.:

Base:  Image-1.0
Layer: Image-1.1

The part after the dot was changed frmo 0 to 1 to indicate
it's a layer. This was simple and worked.
But it did not leave us the freedom to use dotted and dashed
version numbers for images, i.e.: ovirt-nod-ng-1.0-1.0.el6
is a nice RPM-ish name, but it couldn't be used as a layer name.

This wasn't a problem until we started delivering images inside
rpms, because in thatc ase it is very convenient to use the rpm
name (actually the rpm nvr) for the image name.

This patch is now to allow us to use the rpm name as the image
name.
The scheme is now as follows:

Base:  Image-1.0-0.4
Layer: Image-1.0-0.4+1

Instead of modifying the existing release, an extra bit is getting
appended (the layer index), to specify that it's the (ni teh example
above) first layer of the given base.

Change-Id: I5d88ac78434ad0126c66d27264961f94a1ee7bfa
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>