imgbase: Reuse add_layer() for initial layout creation
Previously the low-level function _add_lvm_snapshot was called during
the initial layout creation, but this is not necessary, as the
higher-level function add_layer() can be reused with a small adjustment.

This will reduce the number of code flows for snapshot additions and
shoudl further simplify the codebase.

In addition two signals could be removed.

Change-Id: If588f444dcbcc3d44f5e308fee63ab28c5bf11b1
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>