osupdater: Ensure that /etc is mounted after initialization
Previously /etc was still pointing to the old root, after layout
initialization.
With this patch, imgbase will add a bind mount, which redirects
/etc to <new-lv-of-layer>/etc, which will ensure that all changes
done to the current /etc will be passed on the the new snapshot.

This addresses the issue that changes done to /etc after
initialization will be lost.
Now they will not be lost anymore.

Change-Id: I981b732cfcec132348feac247e521b619471d40c
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>