bootloader: Refactor other entry removal
This patch tries to clean the code flows again.
Previously handling other and valid entries were in the same flow.
This was now changed again, to have a flow for the other entries, and another
flow for the valid entries.
However, they still use the same logic to retrieve the entries.

Change-Id: Ie76710e3ead5aa05dd573563168c111d311067e6
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1346872
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>