osupdater: Don't remove other entries
Reverting this change, because teh risk is, that if the installation fails,
then we can end up with no boot entries at all.

Thus the removal is getting removed from the installation.
It should however - this will be a separate patch - be done either
manually or after the first successful boot.

Change-Id: I1d8c225738a2fd10ede619fae099e804a38430a8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1346872
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>