osupdater: sync up systemd on upgrade
Verify what's in /etc/systemd/system vs the factory, so we can
pick up services which were disabled (iptables, etc).

Added services (from new images, or enablement) are copied anyway
as part of the rsync, but we don't want to delete with rsync,
because additions in the new image would be removed.

Verify it vs /usr/share/factory/etc to see what changed, and pull
out the differences.

Change-Id: I567941e6e246768200302815cb398896740ab2bf
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1380797
Signed-off-by: Ryan Barry <rbarry@redhat.com>