build: Add IQN randomization
Currently the IQN nam eis defined inside the image, this is bad, because
all hosts will share the same IQN name.

With this patch we a) empty the initiator name in post-processing
b) we add a service which is generating a random IQN name during boot
if none is defined.

systemd variable interpolation tries to do its own thing before
bash. Make this a little more complex so it works as intended

Change-Id: I9e3cb2c53d74d6f5d06db7efa180e5528bda63c7
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1393833
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>
(cherry picked from commit 29453a401beb2da368a45d79182099ef66df0769)