update: use the size of the current layer for the new LV
Previously, we iterated over the size of the squashfs contents
and scaled that up. This means that systems after updating could
result in dramatically lower sizes for / (from 30GB -> 6GB, for
example)

Add a .size property to LVs, and use that size directly when
creating layers for updates.

Remove setting .size attributes in tests, which doesn't actually
appear to do anything

Change-Id: I0bb7be571e3e9c10a3bc238e21d297be9835aba4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1376042
Signed-off-by: Ryan Barry <rbarry@redhat.com>