service: copy kernel and initrd to /boot on startup
Some tooling (grub2-mkconfig, libguestfs) does not know how to
handle the kernel and initrd being in subdirectories of /boot.

Copy them to /boot on system startup, and copy them back on
system shutdown (in case of a dracut rebuild)

We can use a service and a trivial bash script to find the kernel,
then infer the initramfs location based on where the kernel is,
in order to copy them back and forth.

Change-Id: I3770e553babda12603cfc2a3cedfa3bf1b8078ea
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1392904
Signed-off-by: Ryan Barry <rbarry@redhat.com>
(cherry picked from commit 213b28fed5963aeaef8f25982e3391088fef776d)