osupdater: copy FIPS files into /boot
dracut-fips doesn't check for the signature file where it is
specified by BOOT_IMAGE. We need to move it ourselves.

In order for this to work, we also need to copy in the kernel
and initrd.

Change-Id: I781d36796329efbbf7c81800f537b41a3debd507
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1414664
Signed-off-by: Ryan Barry <rbarry@redhat.com>
(cherry picked from commit 77b704b44f201587e63051a6b72c1ea6344bba1a)