osupdater: migrate to NIST partitioning
Break the "Volumes" class out of plugins so it's nicer to use

Pull in contextlib.ExitState from python3, so we can easily set an
arbitrary number of running contexts (to bind mount paths).

Check whether or not the given partitions are already mounted, and
create them as volumes if they're not.

This needs to happen before migrate_etc so all of the systemd unit
files and settings can be copied.

Change-Id: Ibe2ec577736369bb33b70a8114ae590f77354fdb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1420068
Signed-off-by: Ryan Barry <rbarry@redhat.com>