osupdater: Fix NIST migration
Add an extra default arg to volume.create() for whether the
volume should be activated immediately.

When changing to NIST layouts, don't activate /tmp, or bad things
happen to umount. This allows us to drop the utils.unmount check.

Additionally, bind mount the new /etc so the new systemd units
are put onto the new FS, as --init will miss this, and we only
want them active on the new layers anyway.

Change-Id: Id74504a29c20a1b7a8b22f0417fa0c89e53f420b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1420068
Signed-off-by: Ryan Barry <rbarry@redhat.com>