osupdater: add new users/groups on update
Previously, imgbased did not pick up nwe users and groups from
the new layer.

Add them (with new IDs if necessary), then add the new IDs to the
list for IDMap() to change ownership of.

Finally, write the content to /etc/passwd and /etc/group on the
new layer once we're finished.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1436172
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1435887

Change-Id: Ibfec6044a0ecdf3cab2b8c5a016aa1ac11013e84
Signed-off-by: Ryan Barry <rbarry@redhat.com>