osupdater: run selinux commands in RPM %post scripts on upgrade
Currently, this only seems to affect ovirt-vmconsole[-host], as
no other packages on the system use fixfiles, semanage, or
semodule in %post. However, we'll try to catch them all.

Add an RpmPackage abstraction, which will parse out the scriptlets
or programs in each section into a dict. Then go through all the
packages on the new image, check if they have %postinstall, and
pull out any lines which have fixfiles, semanage, or semodule.

Then run those in the context of the new image.

Change-Id: I725904565202826430340bfbae7c8e4db60cc206
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1419829
Signed-off-by: Ryan Barry <rbarry@redhat.com>