gc: keeping at least current and new bases
The garbage collector is configured to keep 2 recent base layers, this
change makes sure that the 2 bases to keep are the current layer's base
and the new base that was just installed.

If the GC is configured to keep more than 2 bases, the most recent extra
bases will be kept as well.

Change-Id: I3f98fc5fb5f4e901ac3a5f891ff76d03718bf199
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1433189
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>