general: add a translator between ntp and chrony
Since ntp is finally being removed in 4.2, give us an easy way to
convert existing NTP configurations to chrony.

Most of the options are the same, but not all. Chrony supports
more features than ntp, but we will not try to guess which servers
are pool and which are not.

Here, we will also make the assumption that nodes will not be ntp
servers, so any differences in syntax there can be ignored.

Filter out configuration options chrony doesn't know about which
are present in the default CentOS or Fedora configs (and likely
to be found).

Call this in osupdater.

Change-Id: I35ae296e6264d4992d03b7e84a705f48d3e4ae81
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1503148
Signed-off-by: Ryan Barry <rbarry@redhat.com>
(cherry picked from commit c9838bcd90052cb781aaec910cf7f0b27228bdf3)