osupdater: clean up the files from removed boot dirs
imgbased cleans up /boot/ovirt-* when layers are removed, but any
files contained in that path which have been copied to /boot to
work around platform limitations with FIPS, virt-v2v, or kdump
are not, which can lead to /boot filling up on systems which have
had a large number of updates applied.

When we identify a path in /boot/ovirt-* to be removed, find any files
which were present in it, and remove them from /boot as well.

In addition, track down kdump kernels (which are dropped directly
in /boot) and remove those also.

Change-Id: Ibf198d0a71d22f0030ef2c4f789a2adfa26b0f72
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1533931
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1533871
Signed-off-by: Ryan Barry <rbarry@redhat.com>
(cherry picked from commit 1a10a0ba16bac9e4dbc76e8cdf476edc02ad3d4b)