osupdater: don't copy extraneous files
In trying to remediate the behavior of early versions of 4.0, we
may accidentally copy removed files to the next layer to try to
mimic the original rsync code.

Before moving on to the next pair of layers, take a look at a
dircmp and find any files which should not be present, ensure that
they are not additions (from /usr/share/factory), and remove them
if necessary.

Change-Id: I51ccfa03242e928068e23e5794b381a56c4c287a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1535791
Signed-off-by: Ryan Barry <rbarry@redhat.com>
(cherry picked from commit f1fc848e4fb452847272a44aa14c88ed2ac83d9f)