osupdater: Check filecmp when updating /etc
When comparing /etc to /usr/share/factory/etc during upgrades, it
is possible that a persistent file which becomes a symlink in
later versions will spuriously be copied.

For example:

layer a:
/etc/redhat-access-insights/insights-client.conf (file)
layer b:
/etc/redhat-access-insights/insights-client.conf (symlink)
layer c:
/etc/redhat-access-insights/insights-client.conf (symlink)

The file in /etc is always correct, but imgbased will see that it
matches /usr/share/factory on the first layer, and proceed to
remediate it, even with a dircmp.

When encountering links, shutil.copy2() treats all links as the
same file.

Log it so we know, but don't bomb out of the upgrade.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1575922
Change-Id: I9854d2629575e5a72c37c59e6bb5d01c2489daa8
Signed-off-by: Ryan Barry <rbarry@redhat.com>