osupdater: remove broken symlinks when syncing layers
Previously, an update with the following scenario:

/etc/bar
/etc/foo -> /etc/bar

May remove /etc/bar first, leaving /etc/foo as an orphaned symlink
which fails os.path.isfile(). Add a check for files in our list of
"to remove" which double-checks whether a given file is a broken
symlink, and remove it also.

Additionally, ensure that we only log we are deleting a file when
we are sure it will be removed, after all os.path checks

Change-Id: I311bbecc68f197c144634218e679589ff16a762f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1598781
Signed-off-by: Ryan Barry <rbarry@redhat.com>