bootloader: be more careful when handling grub.cfg
The pre-layer-removed hook calls remove_boot which will remove the boot
entries from grub.  This patch will emit pre-layer-removed before the
actual removal, and after the assertions, otherwise we are left with a
working layer but no boot entry.

Added a check in remove_boot to make sure we are not removing the last
boot entry available.

Backing up grub.cfg before making any changes, and restoring in case the
grubby call failed.

Change-Id: Ic785d2f8338634577fe04cc6b46582a73450d998
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>