Sync existing files in /var
If a file from the new image already exists in /var we skip it, and this
may cause us to skip some files that were updated.  This patch will move
the existing file and copy over (or symlink) the new file.  If the user
wants to rollback, they will need to restore the previous files.

Change-Id: I6cf9695b17e497d4ba00ebf334f2dcc67a1a230b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1643733
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>