osupdater: don't copy the libvirt/qemu cache on upgrades
qemu/libvirt keep a capabilities cache of what the host CPU can
offer, which is checked against every so often in order to ensure
that it's current. However, there's a bug in the implementation,
so new kernel installs/upgrades can result in a host presenting
additional capabilities while the system has a stale cache.

On CentOS, this cache is removed on upgrades. For Node, we can
simply skip syncing it to the new /var and achieve the same effect

Change-Id: I1ceedc2b87cd41980bcf86930245b0d2e6e54ea5
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1674265
Signed-off-by: Ryan Barry <rbarry@redhat.com>
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>