py3: dict.iteritems not available anymore
in python3 dict.iteritems is not avaialble anymore.
dict.items is available in both python2 and python3 but in
python2 it's slower due to implementation not using iterators.
We could have used six.iteritems(dict) to be compatible with both
and keep using iterators but since we don't really need to keep
python2 compatibility, avoiding to add a dependency here.

Change-Id: I7ea3ad97d2dd935e2693139e0f9be36d432ba489
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>