bootsetup: add discard to rootflags
Remounting an xfs filesystem to add the discard option is not supported
as in ext4.  Since / is mounted in the initrd without the discard
option, it can't be remounted later with discard, so we need to use the
boot param rootflags to make sure discard is enabled.

Updated the discard health check to test against the existing mount
point instead of fstab or systemd.

Change-Id: Idf7175ac0a4342b3c067e0104862e407d7700466
Bug-Url: https://bugzilla.redhat.com/1803017
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>