osupdater: Change crypto-policies files permissions
In el8.3, the symlink file in crypto-policies was changed
from %config to %ghost file, but it was kept as a symlink,
and rpm for %ghost files does record 777 permissions of the symlink,
this is causing ssh connection on RHVH with active secure profile to fail
after using rpm --setperms.
Changing the files permissions to 644 as a workaround.

Bug-Url: https://bugzilla.redhat.com/1902646
Signed-off-by: Asaf Rachmani <arachman@redhat.com>
Change-Id: I3b7e864d5e3fbc77273f7cb9bb807e551f1b326e