opsupdater: Do not change UID/GID in remediate_etc
Do not change the UID/GID in remediate_etc,
and trust migrate_etc to do that if needed.

In remediate_etc, the UID/GID of directories are copied from the previous layer
to the new layer when using rsync with `-og` options without considering
the timestamp.
This causes the directories to have the same UID/GID in both layers
even if originally this was not the case.
Because of that, migrate_etc doesn't detect the difference
between the layers and keeps the UID/GID as is.

Bug-Url: https://bugzilla.redhat.com/1909782
Signed-off-by: Asaf Rachmani <arachman@redhat.com>
Change-Id: I89327c9b6a6da73d8161c2b0fdd92bad71115f20