Fixed problems with rpm persistence (added auto-healing)
Added more logic to the imgbase update code, so that it will be able to
heal automatically systems broken by the the RHBZ#1936972

It now includes code that will sync the persisted-rpms to the correct
state, one that was supposed to be there if not the relevant bug.

It thus should allow us to handle properly the following cases which
were previously broken due to the bug and required manual intervention:

- RPM was installed, but then removed
  In that case both the manual deletion of the file from the directory
  and removal from the system was required.

- RPM was installed and then upgraded (once or more)
  In that case the manual removal of the older version(s) were required
  prior to the upgrade, as otherwise it was causing the upgrade to fail.

- RPM was installed, upgraded and then downgraded
  Similar to the previous case. In this case actually one or more newer
  versions of the RPM package had to be removed from the directory,
  leaving only the correct (not the latest) version.

The new code now sync's against the RPM DB and only keeps the currently
installed versions of the packages, removing all non-installed packages
prior to the re-installation.

Change-Id: I227fe77e46b814ad7e4eeb1cb5e41753b12dd593
Signed-off-by: Lev Veyde <lveyde@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1936972