Fixed the improper handling of SUID and SGID files
Changing ownership of SUID and SGID files requires special care,
as due to the security consideration the Linux clears both bits,
even if the change of the ownership is not related to the bit,
i.e. changing just the group of the file will reset SUID bit, if
it's set.

This patch adds the proper handling for these cases, by saving
the access mode for SUID/SGID files and directories, and then
restoring that access mode, after the ownership change was made.

Change-Id: I0c6217234ea22f48186f2815b277170e95665217
Signed-off-by: Lev Veyde <lveyde@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1964490